requirements.txt
# Cloud-ready requirements (no audio hardware dependencies)
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.8.2
pydantic-core==2.20.1
httpx==0.27.2
python-dotenv==1.0.1
certifi
websockets
# Removed: sounddevice, numpy (not needed for cloud deployment)
